backend/pyproject.toml
[project]
name = "second-memory-backend"
version = "0.1.0"
description = "Second Memory – personal content curation app backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    # Database
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "pgvector==0.3.6",
    # Settings
    "pydantic-settings==2.7.0",
    # AI / LLM
    "langchain==0.3.14",
    "langchain-groq==0.2.3",
    "langchain-community==0.3.14",
    "langgraph==0.2.60",
    # Embeddings (Google AI Studio)
    "google-genai>=1.0.0",
    # Metadata extraction
    "yt-dlp>=2025.1.26",
    "trafilatura==2.0.0",
    "beautifulsoup4==4.12.3",
    "requests==2.32.3",
    "lxml==5.3.0",
    # Utilities
    "python-multipart==0.0.20",
    "httpx==0.28.1",
]
